% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ember_risk.R
\name{calc_ember_risk}
\alias{calc_ember_risk}
\title{Calculate ember risk}
\usage{
calc_ember_risk(prop_table, fuels, FFDI)
}
\arguments{
\item{prop_table}{Table of properties with ancillary fields added, output from generate_property_fields()}

\item{fuels}{Fuels table}

\item{FFDI}{Forest Fire Danger Index for calculation}
}
\value{
Table with ember distance and risk index added
}
\description{
Calculate ember risk
}
\examples{
\dontrun{

calc_ember_risk(houses,fuels,50)
}
}
