% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lsm.R
\name{calc_lsm}
\alias{calc_lsm}
\title{calc_lsm}
\usage{
calc_lsm(
  raster_list,
  category = c("hts", "tsf", "tsfc", "fesm", "other"),
  folder_out,
  year = NULL,
  window_n = 3,
  lsm = NULL,
  mask = NULL,
  overwrite = FALSE,
  plot = FALSE
)
}
\arguments{
\item{raster_list}{List of raster files to or a single raster file path to calculate lsm against.}

\item{category}{Type of categorical data to assess. N.B. Designed for these listed categories -- use 'other' to test against different data, but ensure it is pre-processed to satisfaction.}

\item{folder_out}{Output folder to save processed rasters.}

\item{year}{Optional year to calculate, only used as a filter against the raster_list. Assumes year is in the file name.}

\item{window_n}{Window size to use for a moving window.}

\item{lsm}{Vector list of landscape metrics to calculate. If NULL, it calculates every available option from the {landscapemetrics} package.}

\item{mask}{Optional mask to apply after calculating landscape metrics.}

\item{overwrite}{Flag to overwrite existing file or not.}

\item{plot}{Flag to print a plot of the output before writing data to file.}
}
\value{
Raster layer with time since fire classes.
}
\description{
Calculate landscape metrics using a circular moving window.
}
\examples{
\dontrun{

}

}
