% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_MFDI.R
\name{calc_MFDI}
\alias{calc_MFDI}
\title{Calculate Moorland Fire Danger Index}
\usage{
calc_MFDI(rain, temp, RH, wind, age, time, daily = FALSE)
}
\arguments{
\item{rain}{Rainfall amount (mm)}

\item{temp}{Temperature (degrees C)}

\item{RH}{Relative humidity (RH)}

\item{wind}{Wind speed}

\item{age}{Age of fuel (time since fire)}

\item{time}{Vector of time values in minutes or days}

\item{daily}{If time data is daily set to TRUE, otherwise if it is in minutes set to FALSE}
}
\value{
Moorland FDI
}
\description{
Calculate Moorland Fire Danger Index
}
\examples{
calc_MFDI(rain=c(0,20,10,4),
temp=c(12,13,8,9),
RH=c(55,87,90,67),
wind=c(5,8,9,9),
age=c(25,25,25,25),
time=c(1,2,3,4),
daily=TRUE)
}
