% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rad_risk.R
\name{calc_rad_risk}
\alias{calc_rad_risk}
\title{Calculate radiant-heat risk index for properties}
\usage{
calc_rad_risk(prop_table, fuels)
}
\arguments{
\item{prop_table}{The supplemented property table from gen_property_fields}

\item{fuels}{Data frame containing fuels data (type, loads)}
}
\value{
tibble data frame with fields for fuel break width (FBW), fireline intensity (FL) and distance risk index (DRisk)
}
\description{
Calculate radiant-heat risk index for properties
}
\examples{
\dontrun{
output <- calc_rad_risk(houses,fuel)
}
}
