% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_tsf.R
\name{classify_tsf}
\alias{classify_tsf}
\title{classify_tsf}
\usage{
classify_tsf(raster, mask = NULL)
}
\arguments{
\item{raster}{Raster ({terra}).}

\item{mask}{Optional raster mask layer.}
}
\value{
Raster layer with time since fire classes.
}
\description{
Classify raster time since fire into bins
}
\examples{
\dontrun{

}

}
