% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectable.R
\name{collectable}
\alias{collectable}
\title{collectable}
\usage{
collectable(x, qid = "query", format = "tibble")
}
\arguments{
\item{x}{: dataset}

\item{qid}{chr: optional query identifier to assign to query}

\item{format}{chr: output format type, either via \code{data.table}, \code{dtplyr}, or \code{tibble}}
}
\value{
A queried database connection stored in local memory and the time it took to pull into memory.
}
\description{
Wraps a timer around functions that pulls a lazy database query into memory. Note: This function is re-purposed from \code{toddellis/miao}.
}
