% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cols.R
\name{compare_cols}
\alias{compare_cols}
\title{compare_cols}
\usage{
compare_cols(.x = NULL, .y = NULL, .z = NULL, view = TRUE)
}
\arguments{
\item{.x}{EDW-pulled table or query.}

\item{.y}{EDW-pulled table or query.}

\item{.z}{EDW-pulled table or query.}
}
\value{
Dataframe with column names, classes, and source (.x, .y., and/or .z).
}
\description{
Quick method to compare columns and classes between datasets. Primarily intended for EDW data -- comparable functions (e.g., \code{janitor::compare_df_cols}) do not work with Oracle connections.
}
\examples{
\dontrun{
}
}
