% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_append.R
\name{fc_append}
\alias{fc_append}
\title{fc_append}
\usage{
fc_append(x, tbl, db = "firecentre")
}
\arguments{
\item{x}{Dataset to push to database.}

\item{tbl}{Database schema and table name: Must be in \code{schema.table} syntax.}

\item{db}{Database name or path/to/database/object.}
}
\value{
Data exported to database.
}
\description{
Write to the firecentre PostgreSQL database. Adds to an existing database object. Separate from write and remove options to ensure their use is explicit at all times.
}
\examples{
\dontrun{


}

}
