% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_index.R
\name{fc_index}
\alias{fc_index}
\title{fc_index}
\usage{
fc_index(
  idx_name,
  idx_cols = NULL,
  tbl = NULL,
  unique = FALSE,
  drop = FALSE,
  db = "firecentre"
)
}
\arguments{
\item{tbl}{Database schema and table name used when creating indices: Must be in \code{schema.table} syntax.}

\item{db}{Database name or path/to/database/object.}
}
\value{
Data exported to database.
}
\description{
Manage indices in the \code{firecentre} PostgreSQL database. N.B. May be better to manage these from withing pgAdmin.
}
\examples{
\dontrun{


}

}
