% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_remove.R
\name{fc_remove}
\alias{fc_remove}
\title{fc_remove}
\usage{
fc_remove(tbl, db = "firecentre")
}
\arguments{
\item{tbl}{Database schema and table name: Must be in \code{schema.table} syntax.}

\item{db}{Database name or path/to/database/object.}
}
\value{
Data exported to database.
}
\description{
Removes an existing database object from the \code{firecentre} PostgreSQL database. Separate from write and append options to ensure their use is explicit at all times.
}
\examples{
\dontrun{


}

}
