% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_update.R
\name{fc_update}
\alias{fc_update}
\title{fc_update}
\usage{
fc_update(dependencies = NA, force = FALSE)
}
\arguments{
\item{dependencies}{Option to list dependencies to install.}

\item{force}{Option to force install even if currently running.}
}
\value{
Installed packages
}
\description{
Sets or pulls stored GitHub credentials and initiates updates of firekit package.
}
\examples{
\dontrun{
fc_update(force = TRUE)
}
}
