% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_write.R
\name{fc_write}
\alias{fc_write}
\title{fc_write}
\usage{
fc_write(x, tbl, db = "firecentre")
}
\arguments{
\item{x}{Dataset to push to database.}

\item{tbl}{Database schema and table name: Must be in \code{schema.table} syntax.}

\item{db}{Database name or path/to/database/object.}
}
\value{
Data exported to database.
}
\description{
Write to the firecentre PostgreSQL database. Creates a new table or overwrites an existing one. Separate from append and remove options to ensure their use is explicit at all times.
}
\examples{
\dontrun{


}

}
