% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi_curve.R
\name{fhi_curve}
\alias{fhi_curve}
\title{fhi_curve}
\usage{
fhi_curve(raster, step = 10, filter_severity = TRUE)
}
\arguments{
\item{raster}{{terra} SpatRaster reflecting fires (either binary 1/0 or FESM severity data).}

\item{step}{Step to use for internal buffers. User larger values on larger fires, smaller values on smaller fires.}

\item{filter_severity}{Flag to filter fire severity to high and extreme fires \emph{if} the raster input layer is FESM.}
}
\value{
Dataframe prepared for fhi_nls(...)
}
\description{
Calculate a fire heterogeneity index curve for a raster layer with showing fires.
}
\examples{
\dontrun{

}

}
