% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi_nls.R
\name{fhi_nls}
\alias{fhi_nls}
\title{fhi_nls}
\usage{
fhi_nls(x, start_fhi = 0.01, plot = FALSE)
}
\arguments{
\item{x}{Output of \code{r_fhi_curve}.}

\item{start_fhi}{Start parameter for calculating the FHI from the NLS algorithm. Change if processing errors.}

\item{plot}{Flag to plot the outcome curve and the identified FHI.}
}
\value{
Either a ggplot output or a single FHI value.
}
\description{
Run a non-linear least squares algorithm on the output of the \code{r_fhi_curve} function and extract the fire heterogeneity index. Optionally plot the outcome.
}
\examples{
\dontrun{

}

}
