% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_project.R
\name{fire_project}
\alias{fire_project}
\title{fire_project}
\usage{
fire_project(
  quiet = FALSE,
  dir = c("dropbox", "onedrive", "custom"),
  drive = NULL,
  number = TRUE
)
}
\description{
Creates common folder structure for new projects for consistency.
}
\examples{
\dontrun{
fire_project()
}
}
