% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_tbl.R
\name{fire_tbl}
\alias{fire_tbl}
\title{fire_tbl}
\usage{
fire_tbl(
  query = NULL,
  db = "firecentre",
  reset = FALSE,
  writable = FALSE,
  kid = 1,
  ...
)
}
\arguments{
\item{query}{SQL SELECT query or SCHEMA.TABLE EDW reference.}

\item{db}{Database name or path/to/database/object.}

\item{reset}{Logical Denotes whether to reset the connection object (\code{con}). Useful if needing to turn writable on/off or change databases.}

\item{writable}{Logical re: making the connection writeable. Warning: Only one individual can be connected to a database with a writable connection at a time.}

\item{kid}{int : User ID number for accessing a list of stored credentials via \code{keyring}. Should typically be 1 unless multiple credentials are stored on one system.}
}
\value{
Lazy connection or local database dataframe.
}
\description{
Connect to a specified database/table.
}
\examples{
\dontrun{


}

}
