% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firekit_update.R
\name{firekit_update}
\alias{firekit_update}
\title{firekit_update}
\usage{
firekit_update(dependencies = NA, force = FALSE)
}
\arguments{
\item{dependencies}{Option to list dependencies to install.}

\item{force}{Option to force install even if currently running.}
}
\value{
Installed packages
}
\description{
Sets or pulls stored GitHub credentials and initiates updates of firekit package.
}
\examples{
\dontrun{
firekit_update(force = TRUE)
}
}
