% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeless.R
\name{labeless}
\alias{labeless}
\title{labeless}
\usage{
labeless()
}
\value{
ggplot object with no labels.
}
\description{
Remove all text labels in a ggplot object. Standard for reporting in PowerPoint.
}
\examples{
\dontrun{

}
}
