% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs.R
\name{pkgs}
\alias{pkgs}
\alias{package_check}
\title{pkgs}
\usage{
pkgs(packages, quiet = FALSE, repos = c("https://cloud.r-project.org"))

package_check(
  packages,
  quiet = FALSE,
  repos = c("https://cloud.r-project.org")
)
}
\arguments{
\item{packages}{Single character or character vector with package/library names}

\item{quiet}{Logical whether to show which packages were installed, loaded, or skipped, as well as their version numbers.}

\item{repos}{Specified repo web address to attempt to pull uninstalled libraries from.}
}
\value{
Environmental conditions: Installs and/or loads libraries where necessary.
}
\description{
Install and/or load any necessary libraries, or skip if they are already loaded. Note: This function is re-purposed from \code{toddellis/miao}.
}
\examples{
\dontrun{
utastoolkit::pkgs(c('dplyr', 'tidyr', 'utas-analytics/utastoolkit'),
                  quiet = FALSE)
}
}
