% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ember.R
\name{predict_ember}
\alias{predict_ember}
\title{Calculate 95th percentile ember distance from FDI, slope and fuel load}
\usage{
predict_ember(FDI, slope, fuel)
}
\arguments{
\item{FDI}{Forest Fire Danger Index value}

\item{slope}{Slope type, options are "Flat" and "Hilly"}

\item{fuel}{Surface fuel load in t/ha}
}
\value{
Distance in m
}
\description{
Calculate 95th percentile ember distance from FDI, slope and fuel load
}
\examples{
\dontrun{

predict_ember(40,"Flat",15)
}
}
