% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fire.R
\name{theme_fire}
\alias{theme_fire}
\title{theme_fire}
\usage{
theme_fire(
  axis.title = 10,
  axis.text = 7.5,
  strip.text = 7,
  legend.text = 9.25,
  line_col = "#9B9D9C",
  strip.background = "white"
)
}
\arguments{
\item{axis.title}{Text size for axis title.}

\item{axis.text}{Text size for axis text.}

\item{strip.text}{Text size for facet strips.}

\item{legend.text}{Text size for legend.}

\item{line_col}{Color for axis lines and ticks; defaults to dark grey.}
}
\value{
Theme ready for adding to ggplot2 object.
}
\description{
Use Fire Centre-preferred ggplot2 theme. Default sizing tends to work well for Quarto publications.
}
\examples{
tibble::tibble(x = 1:10,
               y = 10:1) |>
 ggplot(aes(x = x, y = y)) +
 geom_point() +
 theme_fire()

}
