% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_curve.R
\name{threshold_curve}
\alias{threshold_curve}
\title{threshold_curve}
\usage{
threshold_curve(
  df,
  x,
  y,
  n_range = 100,
  dir = c(">", "<"),
  extend = NULL,
  rescale = FALSE
)
}
\arguments{
\item{df}{Dataframe}

\item{x}{Variable from which a threshold is identified (x axis)}

\item{y}{Variable used to identify a constraining threshold (y axis)}

\item{n_range}{Number of values to use to blow out the x axis and identify a threshold.}

\item{dir}{Direction along x axis constraining the threshold.}

\item{extend}{Value between 0 and 1 to blow out the x axis (sometimes important if threshold is near the upper or lower limits; this can crash the model)}

\item{rescale}{Whether to rescale the y axis to between 0 and 1.}
}
\value{
Dataframe prepared for threshold_nls(...)
}
\description{
Reframe data for extracting a threshold or exploring causal relationships.
}
\examples{
\dontrun{
  threshold_curve(mtcars, mpg, cyl, n_range = 25, dir = ">", extend = 0.1, rescale = F)
}

}
