% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_nls.R
\name{threshold_nls}
\alias{threshold_nls}
\title{threshold_nls}
\usage{
threshold_nls(df, x, y, dir = ">", plot = FALSE)
}
\arguments{
\item{df}{Dataframe output by threshold_curve.}

\item{x}{Variable from which a threshold is identified (x axis)}

\item{y}{Variable used to identify a constraining threshold (y axis)}

\item{plot}{Binary determining whether to automatically plot the relationships to assess the ouput.}
}
\value{
Output of broom::tidy() on nls models.
}
\description{
Identify and extract potential threshold by comparing two variables.
}
\examples{
\dontrun{
  threshold_curve(mtcars, mpg, cyl, n_range = 25, dir = ">", extend = 0.1, rescale = F) |>
    threshold_nls(mpg, cyl, dir = ">", plot = TRUE)
}

}
