% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_pal.R
\name{utas_pal}
\alias{utas_pal}
\title{fire_pal}
\usage{
utas_pal(n = 8, pal = "new")
}
\arguments{
\item{n}{int : Number of unique color values to use.}

\item{pal}{chr : Internal palette choice, including 'new', 'new2', 'classic', 'blue', 'green', 'grey', 'red', 'nature', 'nature2', 'forest', 'oak', 'ochre', and 'basalt'.}
}
\value{
Vector of color hex codes using shades of UTAS-themed colours.
}
\description{
Create custom-length vector of color hex codes that fall within one of the many Fire Centre-themed colour palettes.
}
\examples{
utas_pal(6, pal = 'grey')
utas_pal(3)

}
