% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rad_risk_FFDI.R
\name{calc_rad_risk_FFDI}
\alias{calc_rad_risk_FFDI}
\title{Calculate radiant-heat risk index for properties for defined FFDI}
\usage{
calc_rad_risk_FFDI(prop_table, fuels, FFDI)
}
\arguments{
\item{prop_table}{The supplemented property table from gen_property_fields}

\item{fuels}{Data frame containing fuels data (type, loads)}

\item{FFDI}{FFDI value to calculate risk at}
}
\value{
tibble data frame with fields for fuel break width (FBW), fireline intensity (FL) and distance risk index (DRisk)
}
\description{
Calculate radiant-heat risk index for properties for defined FFDI
}
\examples{
\dontrun{
output <- calc_rad_risk_FFDI(houses,fuel,50)
}
}
