% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_write.R
\name{fc_write}
\alias{fc_write}
\title{fc_write}
\usage{
fc_write(x, tbl, chunk_size = NULL, db = "firecentre", create = FALSE)
}
\arguments{
\item{x}{Dataset to push to database.}

\item{tbl}{Database schema and table name: Must be in \code{schema.table} syntax.}

\item{chunk_size}{Number of rows to insert into the database at a time.}

\item{db}{Database name or path/to/database/object.}

\item{create}{Logical denoting whether to create a new table or overwrite an existing one.}
}
\value{
Data exported to database.
}
\description{
Write to the firecentre PostgreSQL database. Creates a new table or either overwrites or appends to an existing table.
}
\examples{
\dontrun{


}

}
