% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_pal.R
\name{fire_pal}
\alias{fire_pal}
\title{fire_pal}
\usage{
fire_pal(n = 5, pal = "rfs")
}
\arguments{
\item{n}{int : Number of unique color values to use.}

\item{pal}{chr : Internal palette choice, including 'bw', 'climate', 'climcyc', 'fire', 'grey', 'red', 'rfs', 'sinebow', 'utas', 'utas2'}
}
\value{
Vector of color hex codes using shades of UTAS-themed colours.
}
\description{
Create custom-length vector of color hex codes that fall within one of the many Fire Centre-themed colour palettes.
}
\examples{
\dontrun{
  fire_pal(n = 5, pal = "fire")
}

}
