% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_tbl.R
\name{fire_tbl}
\alias{fire_tbl}
\title{fire_tbl}
\usage{
fire_tbl(query = NULL, db = "FIRE_CENTRE_DB", ...)
}
\arguments{
\item{query}{SQL SELECT query or SCHEMA.TABLE EDW reference.}

\item{con}{Database connection object.}
}
\value{
Lazy connection or local database dataframe.
}
\description{
Connect to a specified database/table.
}
\examples{
\dontrun{


}

}
