% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_tbl.R
\name{fire_tbl}
\alias{fire_tbl}
\title{fire_tbl}
\usage{
fire_tbl(
  query = NULL,
  db = "FIRE_CENTRE_DB",
  writable = FALSE,
  proj_str = FALSE,
  ...
)
}
\arguments{
\item{query}{SQL SELECT query or SCHEMA.TABLE EDW reference.}

\item{db}{Database name or path/to/database/object.}

\item{writable}{Logical re: making the connection writeable. Warning: Only one individual can be connected to a database with a writable connection at a time.}

\item{proj_str}{Logical re: if the firekit::fire_project folder structure is used for the location of the database. Temporary measure until databsae storage is (potentially) standardised.}
}
\value{
Lazy connection or local database dataframe.
}
\description{
Connect to a specified database/table.
}
\examples{
\dontrun{


}

}
